refactor(features): 各 feature 域收敛 index barrel（pr / settings / onboarding）
为三个 feature 域各加 index.ts，暴露跨域公共 API（chat 已有先例）：
- pr: PrPanel / PrEmpty / PrItem / usePullRequests
- settings: SettingsModal + ConnectionForm / LlmProfileForm 等表单（onboarding 复用）
- onboarding: OnboardingWizard + OnboardingResult
外部消费者（App / Sidebar / StatusBar 邻近、onboarding→settings、useBootstrap）改走 barrel。
纪律：域内模块互引仍走相对路径、statusbar chip 走子路径，避免循环依赖 / barrel 杂糅。
common / layout 不加 barrel（扁平基础件 / 仅 App 引，收益低）。无行为变更。
