+++ apps/desktop/src/renderer/src/App.tsx
@@ -4,15 +4,13 @@ import type { PrDiscoveryFilter } from '@meebox/shared';
 import { invoke } from './api';
 import { ChatPane } from './components/features/chat';
 import { MainPane } from './components/layout/MainPane';
-import { PrPanel } from './components/features/pr/PrPanel';
-import { PrEmpty } from './components/features/pr/PrEmpty';
-import { OnboardingWizard } from './components/features/onboarding/OnboardingWizard';
-import { SettingsModal } from './components/features/settings/SettingsModal';
+import { PrPanel, PrEmpty, usePullRequests } from './components/features/pr';
+import { OnboardingWizard } from './components/features/onboarding';
+import { SettingsModal } from './components/features/settings';
 import { Sidebar } from './components/layout/Sidebar';
 import { StatusBar } from './components/layout/StatusBar';
 import { TitleBar } from './components/layout/TitleBar';
 import { useToast } from './hooks/useToast';
-import { usePullRequests } from './components/features/pr/hooks/usePullRequests';
 import { useBootstrap } from './hooks/useBootstrap';
 import { usePanelLayout } from './hooks/usePanelLayout';
 import { useUpdateNotice } from './hooks/useUpdateNotice';
