+++ apps/desktop/src/renderer/src/components/features/onboarding/index.ts
@@ -1,0 +1,3 @@
+// features/onboarding 对外公共 API：首启向导组件 + 其收集结果类型。
+export { OnboardingWizard } from './OnboardingWizard';
+export type { OnboardingResult } from './OnboardingWizard';
