+++ apps/desktop/src/renderer/src/components/features/onboarding/steps/LlmStep.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { LlmProfile } from '@meebox/shared';
-import { LLM_PROVIDERS, LlmProfileForm } from '../../settings/LlmProfileForm';
+import { LLM_PROVIDERS, LlmProfileForm } from '../../settings';
 import { LlmProviderIcon } from '../../../common/LlmProviderIcon';
 
 export function LlmStep({
