+++ apps/desktop/src/renderer/src/components/features/onboarding/steps/PlatformStep.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
 import { invoke } from '../../../../api';
-import { ConnectionForm, type ConnDraft } from '../../settings/ConnectionForm';
+import { ConnectionForm, type ConnDraft } from '../../settings';
 import { PLATFORM_META } from '../../../common/PlatformIcon';
 import { FolderIcon } from '../../../common/icons';
