+++ apps/desktop/src/renderer/src/components/features/pr/index.ts
@@ -1,0 +1,6 @@
+// features/pr 对外公共 API。内部模块（PrHeader / PrTabs / tabs/* 等）相互引用走相对路径，
+// 不经此 barrel，避免循环依赖。状态栏 chip 走 features/pr/statusbar/* 子路径，不并入此处。
+export { PrPanel } from './PrPanel';
+export { PrEmpty } from './PrEmpty';
+export { PrItem } from './PrItem';
+export { usePullRequests } from './hooks/usePullRequests';
