+++ apps/desktop/src/renderer/src/components/layout/Sidebar.tsx
@@ -8,7 +8,7 @@ import type {
 } from '@meebox/shared';
 import { invoke, subscribe } from '../../api';
 import { useChatRunStore } from '../../stores/chat-run-store';
-import { PrItem } from '../features/pr/PrItem';
+import { PrItem } from '../features/pr';
 
 // 'conflict' / 'mergeable' 是按远端 merge 状态跨 localStatus 横切的筛选；'all' 不限定
 type FilterKey = 'all' | LocalPrStatus | 'conflict' | 'mergeable';
